Update mistral-medium prices and context sizes (#10729)
* Update mistral-medium prices and context sizes

While testing the Mistral model, I noticed a discrepancy in the pricing shown on the logs screen. After reviewing the code, I confirmed that the pricing values were incorrect.

This PR corrects the input and output token pricing for the latest Mistral model and adds the newly released mistral-medium-2505 version.

* Adds tool calling flag to mistral-medium

* Adds mistral-medium price updates to the main model price file

* Update model_prices_and_context_window_backup.json

sets mistral medium alias to the old values as it probably points to the old version.

* Update model_prices_and_context_window.json

* Update model_prices_and_context_window_backup.json

* Update model_prices_and_context_window.json